final化。文字列比較を = ではなく、StringUtils.equalsで行うよう修正。findbugsによる指摘箇所を修正

+++ org.insightech.er/src/org/insightech/er/db/impl/hsqldb/HSQLDBTableImportManager.java
@@ -125,7 +125,7 @@ public class HSQLDBTableImportManager extends ImportFromDBManagerBase {
 					if (maxValue.longValue() == Long.MAX_VALUE) {
 						maxValue = null;
 					}
-					if (minValue.intValue() == Long.MIN_VALUE) {
+					if (minValue.longValue() == Long.MIN_VALUE) {
 						minValue = null;
 					}
